ready for initial exp

+++ data/EMGDataset.py
@@ -65,8 +65,8 @@ class EMGDataset(Dataset):
         label.reset_index(drop=True, inplace=True)
 
         # convert to tensor
-        self.data = torch.tensor(data.values)
-        self.label = torch.tensor(label.values)
+        self.data = torch.tensor(data.values, dtype=torch.float32)
+        self.label = torch.tensor(label.values, dtype=torch.float32)
 
     #discritize the data into sequences of length seq_len using torch
     def discritize_data(self):
